Rename pairlist.c to node.c
