+++ src/rlang.h
@@ -10,7 +10,7 @@
 #include "sym.h"
 #include "formula.h"
 #include "env.h"
-#include "pairlist.h"
+#include "node.h"
 #include "sexp.h"
 #include "utils.h"
 #include "vector.h"
